+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
+## [v0.2.6] - 2024-07-19
+### Changed
+- Formatted `CHANGELOG.md` by adding `[]` to tags and spaces between versions.
+- Home page column into 2 columns
+- `custom_theme.dart` to `themes.dart` while simplifying by using color seeds.
+- Extracted widgets of home page to `home_page_widgets.dart`.
 
-## v0.2.5 - 2024-07-17
+### Added
+- Search field to home page which shows searched products
+- Sending notifications when clicking on notification settings.
+- System notification permission pop up if notifications aren't enabled.
+- New icon for notification area.
+- `notifs_helper.dart`.
+- `home_page_widgets.dart`.
+---
+<br>
+
+## [v0.2.5] - 2024-07-18
 ### Changed
 - Parameterized SQL queries in `database.dart`.
 - Changed `features.txt` to `CHANGELOG.md` and `TODOs.md`.
@@ -25,8 +41,10 @@
 
 ### Fixed
 - Orders order change after editing by adding order by to select.
+---
+<br>
 
-## v0.2.4 - 2024-06-20
+## [v0.2.4] - 2024-06-20
 ### Added
 - Small changes and code cleanup.
 - SSL mode switch to database connection.
@@ -34,8 +52,10 @@
 
 ### Fixed
 - Issues with editing items; undo functionality now works.
+---
+<br>
 
-## v0.2.3 - 2024-06-19
+## [v0.2.3] - 2024-06-19
 ### Added
 - Page swipe functionality.
 - Merged translations into a single JSON file and made necessary adjustments.
@@ -44,21 +64,27 @@
 - Disabled context warning in `analysis_options`.
 - Made `edit order` functional with validation.
 - Created an edit dialog with changes made to `order_form`.
+---
+<br>
 
-## v0.2.2 - 2024-06-18
+## [v0.2.2] - 2024-06-18
 ### Added
 - Functionality to the select all button.
 
 ### Fixed
 - An issue with the changed variable name `ip_address`.
+---
+<br>
 
-## v0.2.1 - 2024-06-17
+## [v0.2.1] - 2024-06-17
 ### Added
 - `select all` button (functionality pending).
 - `variables.env` for environment variables.
 - Changed window settings from `window_size` to `window_manager`.
+---
+<br>
 
-## v0.2.0 - 2024-06-17
+## [v0.2.0] - 2024-06-17
 ### Added
 - Cloud server connection (thanks to aiven.io).
 - System defaults for theme and language settings.
@@ -67,8 +93,10 @@
 
 ### Changed
 - User account menus updated.
+---
+<br>
 
-## v0.1.9 - 2024-06-13
+## [v0.1.9] - 2024-06-13
 ### Added
 - Icons on theme settings.
 - Selected rows in `order_list` show up in delete selected button.
@@ -78,8 +106,10 @@
 - Optimized `main` by moving parts to `localization.dart`.
 - Optimized imports.
 - Final localization.
+---
+<br>
 
-## v0.1.8 - 2024-06-13
+## [v0.1.8] - 2024-06-13
 ### Added
 - Localization support, some parts in settings missing due to context.
 - Extracted functions from order list, long press selection, bold headers.
@@ -87,8 +117,10 @@
 ### Changed
 - Reduced gaps between columns in `order_list`.
 - Removed checkboxes from `order_list`.
+---
+<br>
 
-## v0.1.7 - 2024-06-12
+## [v0.1.7] - 2024-06-12
 ### Added
 - About page.
 - Dynamic settings title for each navigation destination.
@@ -100,16 +132,20 @@
 - Centered scaffold titles.
 - Made the vertical divider in nav rail more transparent.
 - Darkened dialog barrier color.
+---
+<br>
 
-## v0.1.6 - 2024-06-11
+## [v0.1.6] - 2024-06-11
 ### Added
 - Automatic table creation with required constraints.
 
 ### Changed
 - Small file structure change, settings, homepage, and navbar under navigation.
 - Error and bug fixes for release.
+---
+<br>
 
-## v0.1.5 - 2024-06-10
+## [v0.1.5] - 2024-06-10
 ### Added
 - Final, hopefully, form validation in a separate file.
 - Handled database client errors in `order_list` and `order_form` files.
@@ -121,14 +157,18 @@
 - Removed dialog windows shadow.
 - Updated database name which fixes connection crash.
 - Commented the components.
+---
+<br>
 
-## v0.1.4 - 2024-06-09
+## [v0.1.4] - 2024-06-09
 ### Changed
 - DB IP address input from settings.
 - Finally removed every single trace of that default purplish accent.
 - Default theme set as system theme while maintaining theme switch functionality.
+---
+<br>
 
-## v0.1.3 - 2024-06-07
+## [v0.1.3] - 2024-06-07
 ### Added
 - Filled homepage with products including hero views.
 - Volume slider to notifications settings.
@@ -136,8 +176,10 @@
 ### Changed
 - Dark light mode switch works.
 - Got rid of default purplish accent by overriding accent colors.
+---
+<br>
 
-## v0.1.2 - 2024-06-07
+## [v0.1.2] - 2024-06-07
 ### Added
 - Delete selected order, IDs will refresh on delete.
 
@@ -145,8 +187,10 @@
 - Fixed icon name under `src/main/AndroidManifest.xml`.
 - Row selection now fully works.
 - Removed delete last order button from order form and the functions of it.
+---
+<br>
 
-## v0.1.1 - 2024-06-06
+## [v0.1.1] - 2024-06-06
 ### Added
 - Placeholder text for other settings pages.
 - Placeholder toggles on notification settings.
@@ -155,13 +199,17 @@
 - Check the type of phone, milk, and egg inputs.
 - Check if one of the fields milk, egg, or other is filled.
 - Selectable rows, working on functionality.
+---
+<br>
 
-## v0.0.2 - 2024-06-05
+## [v0.0.2] - 2024-06-05
 ### Added
 - Set custom color scheme.
 - Show snackbar when none of the fields milk, egg, and other are filled.
+---
+<br>
 
-## v0.0.1 - 2024-05-25
+## [v0.0.1] - 2024-05-25
 ### Added
 - Created the project from Flutter template.
 - Writing and reading orders from db works.